change  the Alumni Satisfaction Rate migration

+++ database/migrations/2026_08_12_114946_add_alumni_engagement_rate_and_portal_registration_to_alumni_satisfaction_rates_table.php
@@ -11,11 +11,11 @@ return new class extends Migration {
     public function up(): void
     {
         Schema::table('alumni_satisfaction_rates', function (Blueprint $table) {
-            $table->decimal('alumni_engagement_rate')
+            $table->integer('alumni_engagement_rate')
                 ->nullable()
                 ->after('satisfaction_rate');
 
-            $table->decimal('portal_registration')
+            $table->integer('portal_registration')
                 ->nullable()
                 ->after('alumni_engagement_rate');
         });
